Rename packages to be singular
