+++ command/copy_paste.go
