+++ command/cursor/on_edit.go
