+++ command/editor_redraw.go
