+++ command/find_regex.go
