+++ command/input/applied_hook.go
