+++ command/input/canceler.go
