+++ command/input/change_hook.go
