+++ command/select_all.go
