+++ command/split.go
