+++ command/view.go
