+++ main.go
@@ -18,13 +18,13 @@ import (
 	"github.com/nelsam/gxui/themes/dark"
 	"github.com/nelsam/vidar/commander"
 	"github.com/nelsam/vidar/commander/bind"
-	"github.com/nelsam/vidar/commands"
-	"github.com/nelsam/vidar/commands/input"
+	"github.com/nelsam/vidar/command"
+	"github.com/nelsam/vidar/command/input"
 	"github.com/nelsam/vidar/controller"
 	"github.com/nelsam/vidar/editor"
 	"github.com/nelsam/vidar/navigator"
 	"github.com/nelsam/vidar/plugin"
-	"github.com/nelsam/vidar/settings"
+	"github.com/nelsam/vidar/setting"
 	"github.com/nelsam/vidar/theme"
 	"github.com/spf13/cobra"
 	"github.com/tmc/fonts"
