+++ navigator/projects.go
@@ -7,7 +7,7 @@ package navigator
 import (
 	"github.com/nelsam/gxui"
 	"github.com/nelsam/vidar/commander/bind"
-	"github.com/nelsam/vidar/settings"
+	"github.com/nelsam/vidar/setting"
 )
 
 type ProjectSetter interface {
