+++ plugin/goimports/goimports.go
@@ -16,7 +16,7 @@ import (
 	"github.com/nelsam/vidar/commander/bind"
 	"github.com/nelsam/vidar/commander/input"
 	"github.com/nelsam/vidar/plugin/status"
-	"github.com/nelsam/vidar/settings"
+	"github.com/nelsam/vidar/setting"
 )
 
 const stdinPathPattern = "<standard input>:"
