+++ plugin/license/header_update.go
@@ -11,7 +11,7 @@ import (
 	"github.com/nelsam/vidar/commander/bind"
 	"github.com/nelsam/vidar/commander/input"
 	"github.com/nelsam/vidar/plugin/status"
-	"github.com/nelsam/vidar/settings"
+	"github.com/nelsam/vidar/setting"
 )
 
 type Projecter interface {
