+++ setting/default_project_windows.go
