+++ setting/legacy.go
