+++ suggestion/suggestion_adapter.go
